backend/common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taskflow-common"
version = "0.1.0"
description = "Common utilities and DB models for TaskFlow"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "bcrypt>=4.1",
    "PyJWT>=2.8",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["taskflow_common"]
